example/requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.8.0
Flask-SQLAlchemy==1.0
Jinja2==2.7
Pillow==2.1.0
SQLAlchemy==0.8.2
Werkzeug==0.9.3
argparse==1.2.1
async==0.6.1
blinker==1.3
distribute==0.6.24
gitdb==0.5.4
itsdangerous==0.22
pilkit==1.1.5
requests==1.2.3
smmap==0.8.2
wsgiref==0.1.2
